lms-dashboard/requirements.txt
# Flask web framework
Flask>=2.0.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.0

# For database connection pooling
psycopg2-pool>=1.3.0

# For data visualization
datatables.net>=1.13.4
chart.js>=4.0.0

# For front-end interactivity and styling
jquery>=3.6.0
bootstrap>=5.3.0
jquery-ui>=1.13.2